Added BE components. Moved cooling related components to generic instead of diesel.
