+++ CCL.Importer/Proxies/Simulation/ElectricSimulationReplacer.cs
@@ -15,6 +15,8 @@ namespace CCL.Importer.Proxies.Simulation
             CreateMap<TractionMotorSetDefinitionProxy.TransitionDefinition, TractionMotorSetDefinition.TransitionDefinition>();
 
             CreateMap<TractionGeneratorDefinitionProxy, TractionGeneratorDefinition>().AutoCacheAndMap();
+
+            CreateMap<BatteryDefinitionProxy, BatteryDefinition>().AutoCacheAndMap();
         }
     }
 }
