+++ CCL.Types/Proxies/Simulation/Electric/BatteryDefinitionProxy.cs
@@ -1,0 +1,41 @@
+﻿using CCL.Types.Proxies.Ports;
+using System.Collections.Generic;
+
+namespace CCL.Types.Proxies.Simulation.Electric
+{
+    public class BatteryDefinitionProxy : SimComponentDefinitionProxy, IHasFuseIdFields
+    {
+        public readonly BatteryChemistry chemistry = BatteryChemistry.LeadAcid;
+
+        public int numSeriesCells = 36;
+        public float internalResistance = 0.005f;
+        public float baseConsumptionMultiplier = 4f;
+
+        [FuseId]
+        public string powerFuseId;
+
+        public override IEnumerable<PortDefinition> ExposedPorts => new[]
+        {
+            new PortDefinition(DVPortType.READONLY_OUT, DVPortValueType.VOLTS, "VOLTAGE"),
+            new PortDefinition(DVPortType.READONLY_OUT, DVPortValueType.VOLTS, "VOLTAGE_NORMALIZED")
+        };
+
+        public override IEnumerable<PortReferenceDefinition> ExposedPortReferences => new[]
+        {
+            new PortReferenceDefinition(DVPortValueType.ELECTRIC_CHARGE, "NORMALIZED_CHARGE", false),
+            new PortReferenceDefinition(DVPortValueType.ELECTRIC_CHARGE, "CHARGE_CONSUMPTION", true),
+            new PortReferenceDefinition(DVPortValueType.POWER, "POWER", false)
+        };
+
+        public IEnumerable<FuseIdField> ExposedFuseIdFields => new[]
+{
+            new FuseIdField(this, nameof(powerFuseId), powerFuseId),
+        };
+    }
+
+    public enum BatteryChemistry
+    {
+        Unknown,
+        LeadAcid
+    }
+}
